+++ src/main/java/duke/Ui.java
@@ -1,8 +1,5 @@
 package duke;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Scanner;
